+++ src/NuclearFactory.cpp
@@ -1,5 +1,12 @@
 #include "NuclearFactory.h"
 
-EnergySource* NuclearFactory::createEnergySource(int output) {
+/**
+ * @brief Create an EnergySource object of type NuclearSource
+ * 
+ * @param output The output of the NuclearSource
+ * 
+ * @return EnergySource* The EnergySource object created
+ */
+EnergySource* NuclearFactory::createEnergySource(float output) {
 	return new NuclearSource(output);
 }
